+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added an ```sql`...`.prepare('query-name')``` method for creating prepared statements. This can
+  sometimes have measurable performance benefits, especially if the query is very complex to parse and plan.
+
+
 ## [0.10.0] - 2021-03-31
 
 ### Changed